Relocate tests
